#3 認証が必要なURLを別コントローラに移植
